moved ClusterVNodeSettings to elections service tests
